Respec costs coin, and a lot of it
§3.0a asks for respec at "a material or currency cost". It was written before a
currency existed, so the fee was charged in Bonus Points — the very resource it
refunds. Spend 3 points to get 12 back reads as odd, and made the trade hard to
judge.

It now costs coin, and every point comes back. The fee is the friction; a partial
refund on top of it would be charging twice, and §3.0a wants a mis-invested build
to be recoverable rather than merely survivable.

5,000c, doubling
----------------
Anchored rather than guessed. A player at respec age — level 15 to 25, gathering
tier 1 to 3 — earns roughly 90c to 1,200c per walking hour, so the first respec
is a few hours' work: a real decision, never a wall.

Doubling (5,000 / 10,000 / 20,000) rather than the old linear step. It prices out
someone treating respec as a free rebuild button while barely touching a player
fixing one early mistake.

§3.0a's "cheap the first time" described a points economy where 1 point was
nearly free. The intent was that a first correction should not be punishing, and
a few hours' walking is not — while a casual rebuild should still cost something.

The affordability check runs before anything is cleared, so a respec the player
cannot pay for does not wipe their tree. Tested, along with the doubling and the
full refund.

The app states the cost plainly and warns that the next one doubles — someone on
their third respec is looking at 20,000c and should not discover that afterwards.

Full suite: 825/825.

Note on the run: a first attempt showed four failures, all IsTransient with
SetUp errors and 1m11s durations. Clearing orphaned containers and re-running gave
5/5 in 6 seconds. That is the container contention already documented in
STATUS.md, not a logic failure — worth checking the duration before believing a
failure here.

+++ GeoSlayer.Domain/DTOs/Progression/Responses/PlayerUpgradesDto.cs
@@ -9,7 +9,13 @@ namespace GeoSlayer.Domain.DTOs.Progression.Responses
         public int BonusPointsAvailable { get; set; }
 
         /// <summary>What the next respec will cost (§3.0a: cheap first, then escalating).</summary>
-        public int RespecCost { get; set; }
+        /// <summary>
+        /// Coin for the next respec (§5D.4).
+        ///
+        /// <para>Was Bonus Points. Charging the fee in the resource being refunded made the
+        /// trade hard to read; coin comes from something a player can go and earn.</para>
+        /// </summary>
+        public long RespecCost { get; set; }
 
         public List<UpgradeDto> Upgrades { get; set; } = [];
     }
